Tests - Extract DBConnection to own class
Extract the `DBConnection` and move to a dedicated file. This also
changes the extracted class to use the existing logic to initialize the
connection to a test DB.
